Added a Hazard class
Ok so I added a Hazard class that inherits from Actor so that it (in
theory) should be able to be lifted by the bubble. Currently, the object
can tell if it is touching something and not touching something, but it
doesn't yet know what kind of object it is touching. I should hopefully
fix that soon!