Added Respawning in Main Scene
I went into the Player Code and built off of Gigi's platforming stuff.
If a Player touches a Hazard, it's set to teleport to a spawn point we
can set in the editor. The Hazard class doesn't really do anything yet,
but it is set up if we want to play with colliding different "hazards"
or objects. I also moved the Hazard material into the right folder. I
think I forgot to do that in lab, but it's fixed now.