Added Spawning
I built upon Gigi's code and added Player respawning if they touch the
Hazard. The spawn point can be defined in editor, so we can move things
around. I accidently pushed this to master and reverted it, so if there
is a bug with anything I readded quickly, let me know!