Title Screen and Level Select
I put in some basic screens, mostly with some placeholder stuff. The
Level Select has a grid layout to store future buttons, which I will
label later, so we can easily edit the layout when we make new levels.
Additionally, there is a reset button that works, but does not currently
follow the player or change position. I'll look into changing that.

It seems like the biggest issue comes from some graphical bug that comes
only when my code is being called/when a  button is pressed. Everything
becomes darker when a button is pressed and I don't know why. Disabling
the light or tweaking the mesh settings for the objects don't seem to
help. I may have to alter how I'm switching screens. For now though, at
least it kind of works.