Reset Button now moves with Camera
The reset button on our test level now changes position with the camera.
I also make the Canvas adjust with screen size in all files and removed
the unneeded button in the Level Select menu.

I still have no idea what is causing the bug that alters the colors of
the objects, but at least I know from testing it is not due to me not
using the SceneManager.