Attempted to Add Pressure Plate Code
I am working on trying to add a pressure plate type button we could use
in puzzles that activates on collision. I was playing around with
manipulating the Wind so that it turns off and on when something is
colliding with the pressure plate, but for some reason it isn't working.
I'll work on some levels in the meantime, and look at this later.