Made a small level
Due to how the wind is used, my pressure plate is useless for affecting
it, even if the object is not active. I'll look into how to fix it
later, but for now I made a simple platforming level. I'll polish it up
later, but for now it seems to work ok. I haven't added the level to the
level select screen, but I plan to put it in after I tweak level select
screen. I notice the dimensions sometimes mess up on certain screen
sizes, so I'll see if I can make it uniform.