Added Lightmaps & How to Play pop-up
I added lightmaps and a how to play button that makes a pop-up appear on
the main menu. Please don't delete the folders for your levels as those
contain the lightmaps without which the level won't look right.

A couple the the levels were a bit buggy, so I tweaked them a bit to
ensure everything will work for tomorrow. Will try to add more tonight
though, see if I can't polish it up a bit.